server/ai/requirements.txt
# Sistema de IA - Compatível com Python 3.13
openai>=1.12.0
telethon>=1.35.0
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.108.0
uvicorn>=0.25.0
pydantic>=2.5.3
motor>=3.4.0
redis>=5.0.1
schedule>=1.2.1
aiofiles>=23.2.1
